+++ .github/sys-tests/testcase-2/test.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+docker compose up -d --wait
+
+diff -w ../../testdata/basic.ics <(curl -s http://localhost:8080/profiles/test)