+++ cmd/ical-relay/main.go
@@ -7,6 +7,9 @@ import (
 	"net/http"
 	"os"
 
+	"github.com/jm-lemmi/ical-relay/datastore"
+	"github.com/jm-lemmi/ical-relay/helpers"
+
 	"github.com/alexflint/go-arg"
 	"github.com/gorilla/mux"
 	log "github.com/sirupsen/logrus"
@@ -18,6 +21,7 @@ var version string // If you are here due to a compile error, run go generate
 
 var configPath string
 var conf Config
+var dataStore datastore.DataStore
 
 var router *mux.Router
 
@@ -29,9 +33,8 @@ func main() {
 		Notifier     string `help:"Run notifier with given ID"`
 		ConfigPath   string `arg:"--config" help:"Configuration path" default:"config.yml"`
 		Verbose      bool   `arg:"-v,--verbose" help:"verbosity level Debug"`
-		Superverbose bool   `arg:"--superverbose" help:"verbosity level Trace"`
-		ImportData   bool   `arg:"--import-data" help:"Import Data from Config into DB"`
-		Ephemeral    bool   `arg:"-e" help:"Enable ephemeral mode. Running only in Memory, no Database needed."`
+		SuperVerbose bool   `arg:"--superverbose" help:"verbosity level Trace"`
+		ImportData   string `arg:"--import-data" help:"Import Data from Data.yml into DB"`
 	}
 	arg.MustParse(&args)
 
@@ -40,7 +43,7 @@ func main() {
 	if args.Verbose {
 		log.SetLevel(log.DebugLevel)
 	}
-	if args.Superverbose {
+	if args.SuperVerbose {
 		log.SetLevel(log.TraceLevel)
 	}
 
@@ -51,7 +54,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if !args.Verbose && !args.Superverbose {
+	if !args.Verbose && !args.SuperVerbose {
 		// only set the level from config, if not set by flags
 		log.SetLevel(conf.Server.LogLevel)
 	}
@@ -59,6 +62,21 @@ func main() {
 	log.Trace("Trace log is enabled") // only shows if Trace is actually enabled
 
 	log.Tracef("%+v\n", conf)
+
+	if !helpers.DirectoryExists(conf.Server.StoragePath + "notifystore/") {
+		log.Info("Creating notifystore directory")
+		err = os.MkdirAll(conf.Server.StoragePath+"notifystore/", 0750)
+		if err != nil {
+			log.Fatalf("Error creating notifystore: %v", err)
+		}
+	}
+	if !helpers.DirectoryExists(conf.Server.StoragePath + "calstore/") {
+		log.Info("Creating calstore directory")
+		err = os.MkdirAll(conf.Server.StoragePath+"calstore/", 0750)
+		if err != nil {
+			log.Fatalf("Error creating calstore: %v", err)
+		}
+	}
 
 	// run notifier if specified
 	if args.Notifier != "" {
@@ -73,31 +91,46 @@ func main() {
 		log.Debug("Server mode.")
 	}
 
-	if !args.Ephemeral {
-		if len(conf.Server.DB.Host) > 0 {
-			// connect to DB
-			connect()
-			log.Tracef("%#v", db)
+	// setup router. Will be configured depending on FULL or LITE mode
+	router = mux.NewRouter()
 
-			if args.ImportData {
-				conf.importToDB()
+	if !conf.Server.LiteMode {
+		// RUNNING FULL MODE
+		log.Debug("Running in full mode.")
+		if conf.Server.DB.Host == "" {
+			log.Fatal("DB configuration missing")
+		}
+
+		// connect to DB
+		datastore.Connect(conf.Server.DB.User, conf.Server.DB.Password, conf.Server.DB.Host, conf.Server.DB.DbName)
+		dataStore = datastore.DatabaseDataStore{}
+
+		if args.ImportData != "" {
+			err := datastore.ImportToDB(args.ImportData) // TODO
+			if err != nil {
+				log.Fatalf("Error importing data: %v", err)
 			}
-		} else {
-			log.Fatal("No database configured. Did you mean to start in ephemeral mode?")
+		}
+
+		// setup routes
+		initHandlersProfile()
+		initHandlersApi()
+
+		if !conf.Server.DisableFrontend {
+			htmlTemplates = template.Must(template.ParseGlob(conf.Server.TemplatePath + "*.html")) // TODO: fail more gracefully than segfault
+
+			initHandlersFrontend()
 		}
 	} else {
-		log.Warn("Running in ephemeral-mode. Changes to the config will not persist!!")
+		log.Warn("Running in lite mode. No changes will be saved.")
+		dataStore, err = datastore.ParseDataFile(conf.Server.StoragePath + "data.yml")
+		if err != nil {
+			log.Fatalf("Error loading data file: %v", err)
+		}
+
+		// setup routes
+		initHandlersProfile()
 	}
-
-	// setup template path
-	htmlTemplates = template.Must(template.ParseGlob(conf.Server.TemplatePath + "*.html"))
-
-	// setup routes
-	router = mux.NewRouter()
-	initHandlers()
-
-	// start cleanup
-	CleanupStartup()
 
 	// start server
 	address := conf.Server.Addr